+++ docs/docs/.vitepress/theme/components/vp-demo/index.vue
@@ -1,22 +1,22 @@
-<script lang="ts" setup name="VpDemo">
+<script setup name="VpDemo">
   import { ref } from 'vue'
 
   /**
    * 是否展示内容
    */
-  const isOpen = ref<boolean>(false)
+  const isOpen = ref(false)
   /**
    * 折叠的 dom 节点
    */
-  const content = ref(null as unknown as HTMLDivElement)
+  const content = ref(null)
 
   /**
    * 点击执行
    */
-  const handleClick = (): void => {
+  const handleClick = () => {
     if (!isOpen.value) {
       content.value.style.height = 'auto'
-      const height: number = content.value.offsetHeight
+      const height = content.value.offsetHeight
       content.value.style.height = '0'
       content.value.offsetHeight
       content.value.style.transition = '0.33s'