+++ packages/fighting-design/ripple/src/props.ts
@@ -25,7 +25,7 @@ export const Props = {
    */
   type: {
     type: String as PropType<FightingType>,
-    default: (): FightingType => 'default',
+    default: (): null => null,
     validator: (val: FightingType): boolean => {
       return (['default', 'primary', 'success', 'danger', 'warning'] as const).includes(val)
     }