+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
 - `f-alert` 组件 `type` 默认值改为 null
+- `f-link` 组件 `type` 默认值改为 null
+- `f-ripple` 组件 `type` 默认值改为 null
+- `f-tag` 组件 `type` 默认值改为 null
 
 ## 0.10.0 (2022-12-11)